feat(bilibili): preserving t parameters of link

+++ src/main/kotlin/moe/kurenai/bot/model/bilibili/VideoInfo.kt
@@ -132,7 +132,7 @@ data class VideoInfo(
         @Serializable
         data class Page(
             val cid: Long,
-            val page: Long,
+            val page: Int,
             val from: String,
             val part: String,
             val duration: Int,
